backend/text-to-gloss/requirements.txt
functions-framework==3.*
google-generativeai==0.3.2